+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+## Unreleased
+
 ## 2.0.11.407
 
 - Added PowerShell Core version support badge.